+++ projects/ui-lib/mapper/component-mappers/header-clear.mapper.ts
@@ -1,4 +1,4 @@
-import { mapImage } from '@lluc_llull/ui-lib/utils';
+import { mapImage, mapButtons } from '@lluc_llull/ui-lib/utils';
 import { mapSocialLinks } from './nav-modal.mapper';
 import { ComponentMapperFn } from './types';
 
@@ -9,5 +9,6 @@ export const mapHeaderClear: ComponentMapperFn = (props, cdn) => {
         lang: props?.lang || 'es',
         navigation: props?.navigation,
         socialItems: mapSocialLinks(props?.navigation),
+        homeLink: mapButtons(props?.buttons)[0],
     };
 };